updated game.js and server.js

+++ server/server.js
@@ -86,7 +86,8 @@ wss.on('connection', (ws) => {
 
     ws.on('close', () => {
         console.log('Client disconnected');
+        if (gameProcess) {
             gameProcess.kill();
-   
+        }
     });
 });
